Handle COM interfaces with colliding method names
Resolve name collisions among methods defined on different interfaces, by
renaming some methods with a disambiguating suffix.

It is common for designers of COM interfaces to define multiple versions of
the same interface, where each new interface is a superset of the previous
version. For examples, see these DirectWrite interfaces:

* [IDWriteTextFormat](https://docs.microsoft.com/en-us/windows/win32/api/dwrite/nn-dwrite-idwritetextformat)
* [IDWriteTextFormat1](https://docs.microsoft.com/en-us/windows/win32/api/dwrite_2/nn-dwrite_2-idwritetextformat1)
* [IDWriteTextFormat2](https://docs.microsoft.com/en-us/windows/win32/api/dwrite_3/nn-dwrite_3-idwritetextformat2)
* [IDWriteTextFormat3](https://docs.microsoft.com/en-us/windows/win32/api/dwrite_3/nn-dwrite_3-idwritetextformat3)

These interfaces define methods that have the same name. When implementing a
COM object in Rust, we need a way to distinguish these functions. The ugly
way is to require that different COM interface definitions avoid using the
same name, but that is clearly not feasible.

This implementation looks for name collisions, and simply renames methods
that participate in name collisions by appending the name of the interface.
If _that_ fails, then we use a numeric suffix. (We could just use the IID
as a suffix, but that's a harsh experience when debugging.)